+++ app/ExternalFiles/Extensions.php
@@ -8,8 +8,6 @@
 namespace ExternalFilesInMediaLibrary\ExternalFiles;
 
 // prevent direct access.
-use ExternalFilesInMediaLibrary\Plugin\Crypt_Base;
-
 defined( 'ABSPATH' ) || exit;
 
 /**
@@ -65,7 +63,7 @@ class Extensions {
 	 *
 	 * @return array<int,Extension_Base>
 	 */
-	private function get_extensions_as_objects(): array {
+	public function get_extensions_as_objects(): array {
 		// collect the list.
 		$list = array();