+++ app/ExternalFiles/ImportDialog.php
@@ -71,10 +71,12 @@ class ImportDialog {
 		add_filter( 'efml_user_settings', array( $this, 'add_user_setting' ), 100 );
 
 		// add user-specific configuration.
-		add_action( 'edit_user_profile', array( $this, 'add_user_settings' ) );
-		add_action( 'show_user_profile', array( $this, 'add_user_settings' ) );
-		add_action( 'personal_options_update', array( $this, 'save_user_settings' ) );
-		add_action( 'edit_user_profile_update', array( $this, 'save_user_settings' ) );
+		if ( $this->is_customization_allowed() ) {
+			add_action( 'edit_user_profile', array( $this, 'add_user_settings' ) );
+			add_action( 'show_user_profile', array( $this, 'add_user_settings' ) );
+			add_action( 'personal_options_update', array( $this, 'save_user_settings' ) );
+			add_action( 'edit_user_profile_update', array( $this, 'save_user_settings' ) );
+		}
 	}
 
 	/**
@@ -375,22 +377,18 @@ class ImportDialog {
 		 */
 		$settings = apply_filters( 'efml_user_settings', $settings );
 
+		// loop through the settings and save them.
 		foreach ( $settings as $name => $setting ) {
 			// get the settings full name.
 			$full_name = 'efml_' . $name;
 
-			// bail if setting does not exist in request.
-			if ( ! isset( $_POST[ $full_name ] ) ) {
-				continue;
-			}
-
 			// get the value from request depending on field type.
 			switch ( $setting['field'] ) {
 				case 'checkbox':
-					$value = absint( $_POST[ $full_name ] );
+					$value = isset( $_POST[ $full_name ] ) ? absint( $_POST[ $full_name ] ) : 0;
 					break;
 				default:
-					$value = sanitize_text_field( wp_unslash( $_POST[ $full_name ] ) );
+					$value = isset( $_POST[ $full_name ] ) ? sanitize_text_field( wp_unslash( $_POST[ $full_name ] ) ) : '';
 			}
 
 			// save it in DB.
@@ -467,12 +465,67 @@ class ImportDialog {
 	public function add_user_setting( array $settings ): array {
 		// add our setting.
 		$settings['hide_dialog'] = array(
-			'label'       => __( 'Do not display this dialog for future imports.', 'external-files-in-media-library' ),
-			'description' => __( 'When the dialog is hidden, the above settings are used for importing external files.', 'external-files-in-media-library' ),
+			'label'       => __( 'Hide dialog', 'external-files-in-media-library' ),
+			'description' => __( 'When the dialog is hidden, the above settings are used for importing external files from any directory archive.', 'external-files-in-media-library' ),
 			'field'       => 'checkbox',
 		);
 
 		// return the settings.
 		return $settings;
+	}
+
+	/**
+	 * Check if actual user could use custom settings for imports.
+	 * This depends on the global setting and (if this is enabled) on its role.
+	 *
+	 * @return bool
+	 */
+	public function is_customization_allowed(): bool {
+		// bail if global setting is disabled.
+		if ( 1 !== absint( get_option( 'eml_user_settings' ) ) ) {
+			return false;
+		}
+
+		// get the list of allowed roles.
+		$roles = get_option( 'eml_user_settings_allowed_roles', array() );
+
+		// bail if roles is not an array.
+		if ( ! is_array( $roles ) ) {
+			return false;
+		}
+
+		// check the given roles.
+		foreach ( $roles as $role ) {
+			// bail if role is not a string.
+			if ( ! is_string( $role ) ) {
+				continue;
+			}
+
+			// check if actual user has this role.
+			if ( Helper::has_current_user_role( $role ) ) {
+				return true;
+			}
+		}
+
+		// return false as user does not have an allowed role.
+		return false;
+	}
+
+	/**
+	 * Return list of enabled extensions from settings.
+	 *
+	 * @return array<int,string>
+	 */
+	public function get_enabled_extensions(): array {
+		// get the value of the setting.
+		$setting = get_option( 'eml_import_extensions', array() );
+
+		// if it is not an array, return an empty one.
+		if ( ! is_array( $setting ) ) {
+			return array();
+		}
+
+		// return the setting.
+		return $setting;
 	}
 }